// frontend/src/app/api-client/model/transactionStatus.ts
/**
 * Consola de Operaciones Transaccionales API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Estado dentro de la máquina de estados.
 */
export type TransactionStatus = 'PENDIENTE' | 'EN_REVISION' | 'APROBADA' | 'RECHAZADA' | 'REVERTIDA';

export const TransactionStatus = {

    Pendiente: 'PENDIENTE' as TransactionStatus,

    EnRevision: 'EN_REVISION' as TransactionStatus,

    Aprobada: 'APROBADA' as TransactionStatus,

    Rechazada: 'RECHAZADA' as TransactionStatus,

    Revertida: 'REVERTIDA' as TransactionStatus
};